// dashboard/src/utils/hasura-api/generated/schemas/remoteSchemaPermissionsStepArgs.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Hasura Remote Schema API
 * API for managing remote schemas in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { RemoteSchemaPermissionsStepArgsDefinition } from './remoteSchemaPermissionsStepArgsDefinition';

export interface RemoteSchemaPermissionsStepArgs {
  /** Name of the remote schema */
  remote_schema: string;
  /** Name of the role */
  role: string;
  /** Permission definition */
  definition?: RemoteSchemaPermissionsStepArgsDefinition;
}